// packages/contract-helpers/src/tx-builder/typechain/TxBuilder.ts
export type ActionStruct = {
  target: PromiseOrValue<string>;
  data: PromiseOrValue<BytesLike>;
};